Switch from `https` to `http` for cache repo clones
Per https://github.com/flowtype/flow-typed/issues/226, people seem to be having
issues when the CLI attempts to clone the flow-typed repo into local cache.

Specifically, the github seems to require authentication to do this clone.

There are some reports that cloning over http does not require said auth, so
this switches things to http to attempt to mitigate the problem.