Fix `flow-typed install` for packages without a 'main'/index.js (#349)
Previously we'd use `resolve()` to find the path to an `npm install`ed
dependency. This usually works, but in some cases a package doesn't have
either an `index.js` file or a `main` entry in its package.json. `react-scripts`
is one such example.

This makes our resolve function resilient to this case.