Drop `nodegit` in favor of shelling out to `git` (#358)
We only interface with node in a few key places and `nodegit` just wasn't
pulling its weight. It takes time to compile on install and I also hit some
issues with `revWalk.fileHistoryWalk` returning no results on Windows for some
reason. We might be able to fix these, but this was just as easy and will
probably simplify git coordination well enough for the future of this project
anyway.

This just adds a new `lib/git.js` module to abstract git operations done via
shell-out. This will keep use slightly more dependency-free and hopefully
minimize cross-platform discrepancies only to those in node itself.