New, re-written install command (`flow-typed new-install`)
The medium-term goal is to update the structure of libdef processing so that
npm-specific libdef processing is kept isolated from "generic" libdef
processing.

This will allow for other kinds of non-npm libdefs to exist and be processed in
the flow-typed repo -- namely things like "core" libdefs, which we'd like to
move into flow-typed and out of Flow's core code.