mysql libdef
First approach - specs are done for public, documented interfaces.

Most of the internal types, like PoolNamespace and various protocol
packets are not exposed for now, since they are not often used and
not even documented.

Also, possible events for event emitters are not properly typechecked -
I am not sure, that there is proper way to describe them.