Extend firebase v4 for other older flow versions
This commit extends the typing for firebase v4 to flow version 0.34.
The versioning of the lib type def is also renamed to v4.x.x instead of
v4.0.0 for better compatibility to future minor versions of firebase v4.

This commit also fixes two small bugs with Query.remove() and
ServerValue.TIMESTAMP()