koa: make `port` argument to `app.listen` optional
For running tests against `koa` apps it's common to not pass in a `port` argument to `listen` as in this example: https://github.com/koajs/examples/blob/master/hello-world/test.js

This will make `koa` let the OS assign a random port. This is important during testing so that concurrent tests don't create port clashes.