Fix regression introduced with moment_v2.3.x
This regression got introduced with #1242, when a seemingly older
version of moment_v2.x.x was forked into moment_v2.3.x.  This definition
was previously fixed by me in #890, and this now re-introduces the same
fix for the forked v2.3.x version, which requires the same method
signature for .utcOffset().