Fix type of redis.setex
"seconds" in "SETEX key seconds value" is expected to be a number and
not a string.
See https://redis.io/commands/setex

In "setex('my-key', '5', 'my-value')", the Redis client for JavaScript
will apparently parse 5 as a number, but allowing both string and number
for the "seconds" argument may not be the best thing to do if we
want to limit side-effects/unexpected behaviors through using strict
types.