Replace `any` by `mixed` in express_v4.x.x.js
Using `any` will effectively disable the type checker anywhere you use the
`req.body` property, which is dangerous. By declaring it as a `mixed` type, it
allows you to use it, but you'll have to make sure that any assumptions you
make about its contents are checked first, which is the way to go.