Add response type to axios libdef (#2006)
This PR adds a separate response type to axios calls, earlier the definitions
used the same infered type for config.data and response.data. This has been mentioned in #1975 and #1539.

But hopefully this wont break backwards compatability since I added the
original type (T) as default for the new response type (R). Also verified by tests.

As of now I have only updated the 0.18.x-version. But the same issue is present in earlier versions as well. If you think I should update the old ones as well I'd be happy to do so.

* Add separate response type to axios

THis commit adds a separate response type to axios calls, earlier the definitions
used the same infered type for config.data and response.data.

But hopefully this wont break backwards compatability since I added the
original type (T) as default for the new response type (R).

* Use AxiosPromise when testing types

* Test that types carries through