lodash/fp: Fix compact, unset, dissoc, dissocPath
lodash compact seems to have gotten fixed at some point since the fp definitions were added, this ports the same fix into fp. Namely: The input array is currently `Array<T>`, but it must be 
 `Array<?T>` for an output of `Array<T>`because the whole point of `compact` is to get rid of nil elements.

`unset/dissoc/dissocPath` returns a `boolean` in normal lodash, reflecting the success of the operation. However `fp` does not mutate the input, so instead it returns an `Object` which is a new object resulting from attempting the operation on the input object. This is simply a case of the original `fp` definition writer not noticing that its return type is different from normal lodash.