fix: react-apollo: don't constrain ignored return type of user-passed functions
@TLadd please don't ever stipulate `void` for the return type of a user-passed function, *except, perhaps, if passing an async function is guaranteed to be a mistake.*

It's perfectly reasonable to pass an async function for many callback options, and constraining the return type to `void` in those cases causes an error on async functions.

If whatever calls the user-defined function doesn't care about its result or synchronous completion, the return type should be `any`.

The use of `| void` also seems way suspiciously prevalent in these types.  Are you sure that in all of those cases the value can be `undefined` but not `null`?