fix order of generic types in MergeWithKey
Flow (at least 0.82+) is producing the following errors on current typing, unless this fix is applied:

```

Error --------------------------------------------------------------------------- flow-typed/npm/ramda_v0.x.x.js:1431:23

Cannot use function type as a type because function type is a value. To get the type of a value use `typeof`.

   1431|     A: { [k: string]: T },
                               ^


Error --------------------------------------------------------------------------- flow-typed/npm/ramda_v0.x.x.js:1432:23

Cannot use function type as a type because function type is a value. To get the type of a value use `typeof`.

   1432|     B: { [k: string]: T },
                               ^


Error --------------------------------------------------------------------------- flow-typed/npm/ramda_v0.x.x.js:1439:36

Cannot use function type as a type because function type is a value. To get the type of a value use `typeof`.

   1439|     (<S: string, A: { [k: string]: T }, B: { [k: string]: T }, T>(
                                            ^


Error --------------------------------------------------------------------------- flow-typed/npm/ramda_v0.x.x.js:1439:59

Cannot use function type as a type because function type is a value. To get the type of a value use `typeof`.

   1439|     (<S: string, A: { [k: string]: T }, B: { [k: string]: T }, T>(
                                                                   ^


Error --------------------------------------------------------------------------- flow-typed/npm/ramda_v0.x.x.js:1442:36

Cannot use function type as a type because function type is a value. To get the type of a value use `typeof`.

   1442|     (<S: string, A: { [k: string]: T }, B: { [k: string]: T }, T>(
                                            ^


Error --------------------------------------------------------------------------- flow-typed/npm/ramda_v0.x.x.js:1442:59

Cannot use function type as a type because function type is a value. To get the type of a value use `typeof`.

   1442|     (<S: string, A: { [k: string]: T }, B: { [k: string]: T }, T>(
                                                                   ^


Error --------------------------------------------------------------------------- flow-typed/npm/ramda_v0.x.x.js:1445:36

Cannot use function type as a type because function type is a value. To get the type of a value use `typeof`.

   1445|     (<S: string, A: { [k: string]: T }, B: { [k: string]: T }, T>(
                                            ^


Error --------------------------------------------------------------------------- flow-typed/npm/ramda_v0.x.x.js:1445:59

Cannot use function type as a type because function type is a value. To get the type of a value use `typeof`.

   1445|     (<S: string, A: { [k: string]: T }, B: { [k: string]: T }, T>(
                                                                  
```