[react-apollo] Extract QueryComponentProps
As a user of `react-apollo`, it's a common use case to want to create your own higher-order `Query` component to specify default options, additional behaviours, etc.

In order to type that higher order component, we need to export the Props type of the `react-apollo` `Query` component, so that a consumer of the library can use it to define their HoC props type by extending it.

This commit extracts and exports a generic `QueryComponentProps<TData, TVariables>` type. It has no functional changes.