enforce 1st param on ramda's is function is a type (Class)
Prior to this change, "is" would accept anything for its type in many of the
Flow versions. Now it only accepts a class (Number, String, etc) which could be
realistically enforced by is.

There's also an auxiliary change of any -> mixed where it applies to "is" and in
many cases more tests for how "is" works as well as negative tests.

As of writing these changes have some broken tests that are unrelated to "is"
and instead seem to be resulting from the problem that
https://github.com/flow-typed/flow-typed/pull/2793 is attempting to fix.