make axios.create return a callable result (#3070)
The object returned from axios.create can be called just like the main axios
export directly. The create function is used to produce default settings in a
non-global way for axios.

The Axios class was marked as callable but only as a static form. I'm not sure
what that was doing, but it didn't work in the case of axios.create, and
removing the static bit didn't break any tests.

The callable signature on AxiosExport needs to remain, despite it extending the
Axios class.

I added a simple callable test to an older axios as well.