fix Ramda ifElse to work with variadic arguments
This is very typical when working with curried functions, particularly from
within Ramda itself. The prior definition of ifElse used to work in type
refinement scenarios (although really, it appears to be coercing to any or
otherwise losing type information).

The original test for ifElse is broken because it relies upon such refinements
or loss of type information.

The tests for ifElse are now expressed in a way that is independent of other
Ramda functions.