[react-navigation] Revamp for react-navigation@v4
1. React Navigation v4's main breaking change is that users now have to import directly from `react-navigation-stack`, `react-navigation-tabs`, and `react-navigation-drawer`. I'm introducing libdefs for all the current versions of those packages. I've also removed the corresponding exports from the `react-navigation` package.
2. React Navigation v4 is also TypeScript-typed in-package, so I went through and compared a lot of types and made a bunch of updates. Hard to see the full diff in this PR, but this should be the largest rewrite since the libdef was initially written.
3. We're getting rid of the Flow libdef in the `react-navigation` repo, and will treat the `flow-typed` versions as canonical going forward.

Note that a whole bunch of copy-pasting had to occur. It's currently not possible to import between libdefs, and no tooling exists for generating multiple libdefs from one source. #2842 tracks this issue. The plan for now is to logically separate the libdefs into sections, so it's easier to tell whether or not a change also should affect other libdefs. I've updated the `CONTRIBUTING.md` guidelines accordingly.