add $ReadOnlyArray support to ramda includes + contains
Ramda's includes (same as contains, but contains is deprecated), would not
respect $ReadOnlyArray. These changes allow for $ReadOnlyArray and also enforce
a strict case when it comes to using strings (only strings may be compared with
strings). Tests have been added to enforce this behavior.

Some tests have been problematic to enforce - namely the all-arity version of
$ReadOnlyArray. It is not well understood why this is happening, but there are
comments noting the expected outcome vs. actual. I don't believe it's a
covariance issue, because 1) the curried form works fine, and 2) string and
number shouldn't be covariant.

I also took the liberty of sorting some of the import lists, because it's very
easy to find/place things when it's alphabetized :)