[wip] add three.js support (#3632)
Add a first pass at Three.js. This libdef is incomplete yet pretty big. There
are no accompanying tests as it is a libdef lifted from another project, and
adding tests has prevented this from entering public view. Therefore this is a
work-in-progress that others might use until it meets the expectations of a
given libdef.