fix ramda complement
Ramda's complement returns a function whose type is the same as the input
function provided. The runtime difference is that the boolean value is flipped
when the function is called. Flow doesn't handle expressing this very well, and
thus doesn't preserve the returned function's type. Until we have some mechanism
to do this, we can write a bunch of arity variants - or generate them, as was
done here. The libdef and its accompanying test are both generated.

If you want to generate higher arity to support or want to revisit the
generation, the original generation was done here:
https://github.com/LoganBarnett/typedef-gen