fix qs 6.5.x and add qs 6.9.x (#3631)
* fix qs 6.5.x and add qs 6.9.x

The qs libdefs used the only Function, Object, and sometimes any types - some of
which cause issues in newer versions of Flow and/or flow-typed. The old qs 6.5.x
tests have had the bare minimum done to get the tests passing again. Most of the
work was done against a new (copied) set of libdefs for 6.9.x. I did not know
what versions added which arguments, or if certain arguments were omitted on the
original 6.5.x, and perusing the documentation for qs did not reveal this in a
timely manner. It seemed to make sense make a clean break to 6.9.x. The 6.9.x
version of qs should be entirely backwards compatible with 6.5.x, since only
new, optional arguments have been added.

* fix qs linting issues re: implicit inexact objects