refactor: clean up fs.readFile usage (#3966)
The way `fs.readFile` was being used looked like a code smell, and sure enough, the way it was being mocked would cause surprise errors for anyone who adds idiomatic `fs-extra` calls in new or updated tests.  Pass encoding (`'utf8'`, the default) to get `readFile` to resolve to a string instead of a buffer, and use `readJson` where applicable.  Update fs mocks to behave a bit more like `fs-extra` in this regard (though I'm not sure it would be possible to support all options, but we can at least support encoding string as second argument).