[webpack] Fix import errors (#3995)
After updating to 0.143.0 new errors in lib defs are being highlighted. We should not be importing outside of a module declaration.

- Links to documentation: https://github.com/webpack/webpack
- Link to GitHub or NPM: https://github.com/webpack/webpack
- Type of contribution: fix

Other notes: `OutgoingMessage` doesn't exist in `http` module. Should probably be added in flow but that's probably a different problem to solve. This at least doesn't break the type and keeps it as it was before which would have exposed `any`