[cli] Lock semver to fix install command (#4018)
<!--- # Please remember to use `describe` and `it`in the tests! see https://github.com/flow-typed/flow-typed/blob/master/CONTRIBUTING.md for details. --->
Fixes https://github.com/flow-typed/flow-typed/issues/3988

After deep digging the issue is actually related to the semver package. Versions `7.3.3` and `7.3.4` which are currently latest both break our tests in the following screenshot. We use loose `^` but have a lock file which only tests against current lock which is version `7.3.2`. I've now locked the version but I think we should consider removing the lock file completely as this is a published package and we cannot guarantee the version that will be installed on a users machine matches what's in our yarn.lock (would appreciate feedback on this matter). 

<img width="1012" alt="Screen Shot 2021-02-10 at 5 59 42 pm" src="https://user-images.githubusercontent.com/12436524/107476718-855d9280-6bca-11eb-9846-0af1ed6ae17a.png">


- Links to documentation: N/A
- Link to GitHub or NPM: N/A
- Type of contribution: fix

Other notes: @GAntoine please make sure this is merged before you do release, it's one of the biggest blockers for new people installing flow-typed