Annotate return types for various lodash array functions with Maybe
Functions in `lodash` that operate on `Array<T>` to return `T` should actually use a `?T` for the return type to account for `undefined`, which is given with an empty array as input.

Grepping in the latest source[1] for `: undefined` and narrowing that to `return` sites for exposed functions (tagged as `@memberOf _`) shows the following functions:

- `head`
- `last`
- `nth`
- `max`
- `maxBy`
- `min`
- `minBy`

In addition to the above, `first` is an alias of `head`.

Open questions:

- These functions may not be comprehensive. Unfortunately, the documentation only makes clear that some of these return `undefined`, while omitting this fact for others.
- It's likely that earlier versions of `flow` and previous versions of `lodash` should be updated as well, but as this is the interface I am most exposed to I have started here in case I would need further context to do that.

1: https://github.com/lodash/lodash/blob/4.17.15/lodash.js