webpack 5 (#4047)
- Links to documentation: https://github.com/webpack/webpack/releases/tag/v5.0.0
- Link to GitHub or NPM: https://www.npmjs.com/package/webpack
- Type of contribution: new definition

Other notes: This for now is a straight copy from webpack-4. Generally when new major versions come out it's easier for us to just port the old version to the new one if there are no easily noticeable differences so we can get the version into users hands faster and slowly tweak the defs afterwards.