[cli] install will also install lib defs of dependencies that ship with flow type (#4143)
Closes #4137

- Links to documentation: 
- Link to GitHub or NPM: 
- Type of contribution: fix

### Edit

Example use case:
In my company we flowtype our internal libraries that we then share across codebases. When those libs are built they may have dependencies such as `uuid` or `redux`. While building them, they would need to depend on flow-typed.

When these libs are shipped (lets call one of them `@company/component`) it transpiles into regular js with corresponding `flow.js` files.

Now in my project I want to install `@company/component` and when flow analyses the codebase it would throw uuid module not found errors. This is because `@company/component` needs the definitions for `uuid` but I previously never had it in my `flow-typed/npm` dir.

Ok no worries, I will run `flow-typed install` and everything will be solved. Except because I don't use `uuid` but my dependency does, type defs for uuid will never be installed. I could manually install it with `flow-typed install uuid` but that throws and says I need to specify a version, so now I need to dig into node_modules find the dependency and know that I need to run `flow-typed install uuid@8`.

This is a pretty confusing and not noobie friendly. So with this change, as we parse through all npm libraries that are internally typed and aren't typed through flow-typed we will search their inner dependencies and if it's not already a dependency of the root project it will install their dependencies lib defs too.