[flow|environment] Actually fix environment tests
Summary:
This is my fault, but any time we break out definitions from an existing
module, we need to list the new dependencies alongside the original module. My
previous commits only tested the changed files, but that isn't sufficient for
environment changes.

This fixes all the environment tests, with the exception of `react.js`, which
complains that we are overwriting an existing definition within flow itself.

Test Plan:
- `node cli/dist/cli.js run-tests --environmentOnly`