[flow|environment] Add web-animations spec (#4624)
* [flow-typed] Add environmentOnly option to run-tests

Summary:
It would be helpful to have an option to only run the `environments` definition
test, so this adds one. I don't know if there's a better way to detect
"environment" tests outside of inspecting the path, but this seems to work well
enough for local tests.

Test Plan:
- `node cli/dist/cli.js run-tests --environmentOnly`

* [flow|environment] Actually fix environment tests

Summary:
This is my fault, but any time we break out definitions from an existing
module, we need to list the new dependencies alongside the original module. My
previous commits only tested the changed files, but that isn't sufficient for
environment changes.

This fixes all the environment tests, with the exception of `react.js`, which
complains that we are overwriting an existing definition within flow itself.

Test Plan:
- `node cli/dist/cli.js run-tests --environmentOnly`

* [flow|environment] Add web-animations spec

Summary:
Adds the Web Animations 1 spec: https://www.w3.org/TR/web-animations-1/

Test Plan:
- `node cli/dist/cli.js run-tests --environmentOnly`